requirements.txt
Django==4.2.7
openpyxl==3.1.2
xlsxwriter==3.1.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
